The system attribute in nixpkgs changed because the alias system was dropped. stdenv.hostPlatform.system makes it clear which system is being used (important for cross-compilation) Using stdenv.hostPlatform.system means that:
you’re deliberately choosing the “host” platfom, which is relevant to cross-compilation
you’re picking only system from that (roughly speaking)
The deprecated system alias alone wasn’t that specific (as it not exactly meant “what system I’m on”) so it was deprecated

More info
In case you didn’t manage to fix this, it may be possible that some flake in your system uses the old alias. For that to be fixed you need to contact the flake maintainer to fix this (or create a PR yourself).
To figure out what flake may be causing this, append --option abort-on-warn --show-trace in your nixos rebuild command (if you aren’t using nixos-rebuild maybe try --abort-on-warn --show-trace instead).

See https://discourse.nixos.org/t/how-to-fix-evaluation-warning-system-has-been-renamed-to-replaced-by-stdenv-hostplatform-system/72120